+++ CHANGELOG.md
@@ -4,3 +4,12 @@
 
 ## 0.0.2
 * Fixed issue with movie endpoint throwing a null error
+
+## 0.0.3
+* Added FromMap to MovieFilters
+
+## 0.0.4
+* Added DateTime Parsing of Date Strings
+
+## 0.0.5 
+* Adding GetMovieDetails Function that sets the movie runtime for a movie passed into it